+++ src/components/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren } from 'react';
 import { Footer } from './Footer';
 import { Navbar } from './Navbar';
-import styles from 'src/styles/Layout.module.css';
+import styles from 'src/styles/layout/Layout.module.css';
 
 export const Layout: React.FC<PropsWithChildren> = ({ children }) => {
   return (
