+++ src/components/layout/Navbar.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import { RxHamburgerMenu } from 'react-icons/rx';
 import useToggle from 'src/hooks/useToggle';
-import styles from 'src/styles/Navbar.module.css';
+import styles from 'src/styles/layout/Navbar.module.css';
 import { ToggleMenu } from './ToggleMenu';
 import { Searchbar } from '../ui/Searchbar';
 import { Avatar } from '../ui/Avatar';
