+++ src/components/layout/ToggleMenu.tsx
@@ -1,5 +1,5 @@
 import Link from 'next/link';
-import styles from 'src/styles/ToggleMenu.module.css';
+import styles from 'src/styles/layout/ToggleMenu.module.css';
 import { Avatar } from '../ui/Avatar';
 
 export const ToggleMenu: React.FC<{ toggleAnimation: boolean }> = ({
