+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import styles from 'src/styles/Home.module.css';
+import styles from 'src/styles/Home/Home.module.css';
 
 export default function Home() {
   return (
